requirements.txt
Flask
google-cloud-documentai
google-auth
google-api-core
